+++ src/main/scala/com/jcdecaux/datacorp/spark/internal/HasType.scala
@@ -1,0 +1,9 @@
+package com.jcdecaux.datacorp.spark.internal
+
+import scala.reflect.runtime
+
+trait HasType {
+
+  val runtimeType: runtime.universe.Type
+
+}