+++ src/main/scala/com/jcdecaux/datacorp/spark/internal/Identifiable.scala
@@ -4,10 +4,11 @@ import java.util.UUID
 
 trait Identifiable {
 
-  private[spark] val _uuid: String = UUID.randomUUID.toString
-  private[spark] val _name: String = getClass.getCanonicalName
+  private[this] val _uuid: UUID = UUID.randomUUID
 
-  def getUUID: String = _uuid
+  private[this] val _name: String = getClass.getCanonicalName
+
+  def getUUID: UUID = _uuid
 
   def getCanonicalName: String = _name