+++ src/main/scala/com/jcdecaux/datacorp/spark/workflow/DispatchManager.scala
@@ -1,9 +1,9 @@
 package com.jcdecaux.datacorp.spark.workflow
 
 import com.jcdecaux.datacorp.spark.annotation.InterfaceStability
-import com.jcdecaux.datacorp.spark.exception.AlreadyExistsException
-import com.jcdecaux.datacorp.spark.internal.{DeliverySetterMetadata, Logging}
-import com.jcdecaux.datacorp.spark.transformation.{Deliverable, Factory}
+import com.jcdecaux.datacorp.spark.exception.{AlreadyExistsException, InvalidDeliveryException}
+import com.jcdecaux.datacorp.spark.internal.{HasUUIDRegistry, Logging}
+import com.jcdecaux.datacorp.spark.transformation.{Deliverable, Factory, FactoryDeliveryMetadata}
 
 import scala.collection.mutable.ArrayBuffer
 import scala.reflect.runtime.{universe => ru}
@@ -20,15 +20,24 @@ import scala.reflect.runtime.{universe => ru}
   * </ul>
   */
 @InterfaceStability.Evolving
-private[spark] class DispatchManager extends Logging {
+private[spark] class DispatchManager extends Logging with HasUUIDRegistry {
 
-  private[this] val deliveryRegister: scala.collection.mutable.HashSet[String] = scala.collection.mutable.HashSet()
   private[workflow] val deliveries: ArrayBuffer[Deliverable[_]] = ArrayBuffer()
 
+  /**
+    * Add a new Deliverable object into DispatchManager's delivery pool. <br>
+    *
+    * <b>Attention:</b> Dispatch manager only guarantee that each deliverable object has different UUID.
+    * In other word, use cannot call twice <code>setDelivery</code> with the same argument. However, it doesn't
+    * check if multiple deliverables have the same data type with the same producer and the same consumer
+    *
+    * @param v : Deliverable
+    * @return this DispatchManager
+    */
   private[workflow] def setDelivery(v: Deliverable[_]): this.type = {
-    log.debug(s"Add new delivery of type: ${v.payloadType}")
+    log.debug(s"Add new delivery: ${v.payloadType}. Producer: ${v.producer}")
 
-    if (deliveryRegister.add(v.getUUID)) {
+    if (registerNewItem(v)) {
       deliveries.append(v)
     } else {
       throw new AlreadyExistsException(s"The current deliverable ${v.getUUID} already exists")
@@ -42,7 +51,10 @@ private[spark] class DispatchManager extends Logging {
     * @param deliveryType runtime type
     * @return
     */
-  private[this] def getDelivery(deliveryType: ru.Type, consumer: Class[_], producer: Class[_]): Option[Deliverable[_]] = {
+  @throws[InvalidDeliveryException]("find multiple matched deliveries")
+  private[this] def getDelivery(deliveryType: ru.Type,
+                                consumer: Class[_ <: Factory[_]],
+                                producer: Class[_ <: Factory[_]]): Option[Deliverable[_]] = {
 
     val availableDeliverable = findDeliverableByType(deliveryType)
 
@@ -66,8 +78,20 @@ private[spark] class DispatchManager extends Logging {
             Some(matchedByProducer.head)
           case _ =>
             log.info("Multiple Deliverable with same type and same producer were received. Try matching by consumer")
-            matchedByProducer.filter(_.consumer.nonEmpty).find(_.consumer.contains(consumer))
+            val matchedByConsumer = matchedByProducer
+              .filter(_.consumer.nonEmpty)
+              .filter(_.consumer.contains(consumer))
 
+            matchedByConsumer.length match {
+              case 0 =>
+                log.warn("No deliverable available")
+                None
+              case 1 =>
+                log.debug("Find Deliverable")
+                Some(matchedByConsumer.head)
+              case _ =>
+                throw new InvalidDeliveryException("Find multiple deliveries having the same producer and consumer")
+            }
         }
     }
   }
@@ -78,7 +102,9 @@ private[spark] class DispatchManager extends Logging {
     * @param deliveryType type of data
     * @return
     */
-  private[workflow] def findDeliverableByType(deliveryType: ru.Type): Array[Deliverable[_]] = deliveries.filter(_ == deliveryType).toArray
+  private[workflow] def findDeliverableByType(deliveryType: ru.Type): Array[Deliverable[_]] = {
+    deliveries.filter(_.hasSamePayloadType(deliveryType)).toArray
+  }
 
   /**
     * Collect a [[Deliverable]] from a [[Factory]]
@@ -88,6 +114,15 @@ private[spark] class DispatchManager extends Logging {
     */
   private[workflow] def collectDeliverable(factory: Factory[_]): this.type = setDelivery(factory.getDelivery)
 
+  private[workflow] def dispatch(factory: Factory[_]): this.type = {
+    val setters = FactoryDeliveryMetadata
+      .builder()
+      .setFactory(factory)
+      .getOrCreate()
+
+    dispatch(factory, setters)
+  }
+
   /**
     * Dispatch the right deliverable object to the corresponding methods
     * (denoted by the @Delivery annotation) of a factory
@@ -95,42 +130,39 @@ private[spark] class DispatchManager extends Logging {
     * @param factory target factory
     * @return
     */
-  private[workflow] def dispatch(factory: Factory[_]): this.type = {
-
-    DeliverySetterMetadata.builder()
-      .setClass(factory.getClass)
-      .getOrCreate()
-      .foreach({
-        deliveryInput =>
+  @throws[NoSuchElementException]("Cannot find any matched delivery")
+  @throws[InvalidDeliveryException]("Find multiple matched deliveries")
+  private[workflow] def dispatch(factory: Factory[_], setters: Iterable[FactoryDeliveryMetadata]): this.type = {
+    setters
+      .foreach {
+        setterMethod =>
           // Loop through the type of all arguments of a method and get the Deliverable that correspond to the type
-          val args = deliveryInput.argTypes
-            .map({
-              argsType =>
-                log.debug(s"Dispatch $argsType by calling ${factory.getClass.getCanonicalName}.${deliveryInput.methodName}")
+          val args = setterMethod.argTypes.map {
+            argType =>
+              log.debug(s"Dispatch $argType by calling ${factory.getClass.getCanonicalName}.${setterMethod.name}")
 
-                getDelivery(
-                  deliveryType = argsType,
-                  consumer = factory.getClass,
-                  producer = deliveryInput.producer
-                ) match {
-                  case Some(delivery) => delivery
-                  case _ =>
-                    if (!deliveryInput.optional) {
-                      throw new NoSuchElementException(s"Can not find type $argsType from DispatchManager")
-                    } else {
-                      Deliverable.empty()
-                    }
-                }
-            })
+              getDelivery(
+                deliveryType = argType,
+                consumer = factory.getClass,
+                producer = setterMethod.producer
+              ) match {
+                case Some(delivery) => delivery
+                case _ =>
+                  if (!setterMethod.optional) {
+                    throw new NoSuchElementException(s"Can not find type $argType from DispatchManager")
+                  } else {
+                    Deliverable.empty()
+                  }
+              }
+          }
 
           if (args.exists(_.isEmpty)) {
-            log.warn(s"No deliverable was found for optional input ${deliveryInput.methodName}")
+            log.warn(s"No deliverable was found for optional input ${setterMethod.name}")
           } else {
-            val setterMethod = factory.getClass.getMethod(deliveryInput.methodName, args.map(_.classInfo): _*)
-            setterMethod.invoke(factory, args.map(_.get.asInstanceOf[Object]): _*)
+            val factorySetter = factory.getClass.getMethod(setterMethod.name, args.map(_.classInfo): _*)
+            factorySetter.invoke(factory, args.map(_.get.asInstanceOf[Object]): _*)
           }
-      })
-
+      }
     this
   }