+++ src/main/scala/com/jcdecaux/datacorp/spark/workflow/Flow.scala
@@ -1,6 +1,16 @@
 package com.jcdecaux.datacorp.spark.workflow
 
-private[workflow] case class Flow(payload: String, from: Node, to: Node, stage: Int) {
+import scala.reflect.runtime
+
+/**
+  * Flow is a representation of the data transfer in a Pipeline.
+  *
+  * @param payload type of the transferred data
+  * @param from    origin node of the transfer
+  * @param to      destination node of the transfer
+  * @param stage   stage where the transfer occurs
+  */
+private[workflow] case class Flow(payload: runtime.universe.Type, from: Node, to: Node, stage: Int) {
   def describe(): Unit = {
     println("Flow")
     println(s"Stage     : $stage")