+++ src/test/scala/com/jcdecaux/datacorp/spark/workflow/PipelineSuite.scala
@@ -200,7 +200,7 @@ class PipelineSuite extends FunSuite {
     // Check inspector
     assert(pipeline.pipelineInspector.nodes.size === 4)
     assert(pipeline.pipelineInspector.nodes.find(_.getPrettyName === "Container2Factory").get.input.length === 2)
-    assert(pipeline.pipelineInspector.flows.size === 3)
+    assert(pipeline.pipelineInspector.flows.size === 4)
 
   }
 
@@ -241,9 +241,50 @@ class PipelineSuite extends FunSuite {
     // Check inspector
     assert(pipeline.pipelineInspector.nodes.size === 3)
     assert(pipeline.pipelineInspector.nodes.find(_.getPrettyName === "DatasetFactory2").get.input.length === 3)
-    assert(pipeline.pipelineInspector.flows.size === 3)
+    assert(pipeline.pipelineInspector.flows.size === 5)
 
 
+  }
+
+  test("Test Pipeline creation with default primary constructor") {
+    val spark: SparkSession = new SparkSessionBuilder("test").setEnv("local").getOrCreate()
+    import spark.implicits._
+
+    val ds2: Dataset[Product2] = Seq(
+      Product2("id_of_product1", "c2"),
+      Product2("pd1", "c2")
+    ).toDS
+
+    val pipeline = new Pipeline
+
+    pipeline
+      .setInput(new Deliverable[String]("wrong_id_of_product1"))
+      .setInput[String]("id_of_product1", classOf[ProductFactory])
+      .setInput(ds2)
+      .addStage(classOf[ProductFactory])
+      .addStage(classOf[DatasetFactory], spark)
+      .addStage(classOf[DatasetFactory2], spark)
+      .describe()
+      .run()
+
+    pipeline.get().asInstanceOf[Dataset[Product2]].show()
+    assert(pipeline.get().asInstanceOf[Dataset[Product2]].count() === 2)
+    assert(pipeline.get().asInstanceOf[Dataset[Product2]].filter($"x" === "pd1").count() === 1)
+    assert(pipeline.get().asInstanceOf[Dataset[Product2]].filter($"x" === "id_of_product1").count() === 1)
+    assert(
+      pipeline.get().asInstanceOf[Dataset[Product2]].filter($"x" === "id_of_product1").collect().head ===
+        Product2("id_of_product1", "c2")
+    )
+
+    // Check inspector
+    assert(pipeline.pipelineInspector.nodes.size === 3)
+    assert(pipeline.pipelineInspector.nodes.find(_.getPrettyName === "DatasetFactory2").get.input.length === 3)
+    assert(pipeline.pipelineInspector.flows.size === 5)
+
+    assertThrows[IllegalArgumentException](
+      pipeline.addStage(classOf[DatasetFactory], spark, "sqdfsd"),
+      "IllegalArgumentException should be thrown as the number of constructor argument is wrong"
+    )
   }
 
   test("Pipeline exceptions") {
@@ -312,6 +353,6 @@ class PipelineSuite extends FunSuite {
     assert(pipeline.pipelineInspector.nodes.size === 4)
     assert(pipeline.pipelineInspector.nodes.find(_.getPrettyName === "DatasetFactory2").get.input.length === 3)
     assert(pipeline.pipelineInspector.nodes.find(_.getPrettyName === "DatasetFactory3").get.input.length === 3)
-    assert(pipeline.pipelineInspector.flows.size === 5)
+    assert(pipeline.pipelineInspector.flows.size === 7)
   }
 }