+++ src/test/scala/com/jcdecaux/datacorp/spark/workflow/StepSuite.scala
@@ -1,8 +1,0 @@
-package com.jcdecaux.datacorp.spark.workflow
-
-import org.scalatest.FunSuite
-
-class StepSuite extends FunSuite {
-
-
-}